+++ CHANGELOG.md
@@ -181,3 +181,10 @@ Added:
 ## 0.0.35
 
 - Fixed enableAccuracyEngine() for null accuracy value.
+
+## 0.0.36
+
+- Updated native Roam SDK versions. Android v0.0.26 and iOS v0.0.27
+- Added tracking config methods `Roam.setTrackingConfig()`, `Roam.getTrackingConfig()` and `Roam.setTrackingConfig()`
+- Fixed location tracking crash when location permission is changed for iOS.
+- Fixed blue bar for custom tracking in iOS.